requirements-dev.txt
# Development dependencies for DiaBloS
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.0.0
pytest-qt>=4.2.0
pytest-cov>=4.0.0

# Code quality
pylint>=2.15.0
black>=22.0.0
mypy>=1.0.0

# Type stubs
PyQt5-stubs>=5.15.0
